move SiteOffListener to own Class. refs #1698 This also corrects a non-working admin login when site was closed.
