+++ src/lib/Zikula/Bundle/CoreBundle/Resources/config/core.xml
@@ -106,6 +106,10 @@
             <argument type="service" id="logger" on-invalid="ignore" />
         </service>
 
+        <service id="zikula.site_off_listener" class="%zikula.site_off_listener.class%">
+            <tag name="kernel.event_subscriber" />
+        </service>
+
         <service id="zikula.theme_listener" class="%zikula.theme_listener.class%">
             <tag name="kernel.event_subscriber" />
             <tag name="monolog.logger" channel="request" />
